Title: 특정 문자로 시작하는 문자열, Time: 132ms, Memory: 10MB, Status: Passed - CodeTree

+++ 241207/특정 문자로 시작하는 문자열/strings-that-start-with-a-specific-character.java
@@ -21,7 +21,7 @@ public class Main {
             }
         }
 
-        double lengthAvg = lengthSum / cnt;
+        double lengthAvg = (double)lengthSum / cnt;
         System.out.printf("%d %.2f", cnt, lengthAvg);
     }
 }
